+++ packages/core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unpublished
 
 - Feature: Additionally export `PolarCore` type.
+- Fix: The GFI's new flag `userInteraction` on the close interaction is forwarded. This is required for a fix in the GFI plugin.
 
 ## 1.4.0